requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.3.1
    # via django
attrs==20.3.0
    # via fiona
branca==0.4.2
    # via folium
certifi==2020.12.5
    # via
    #   fiona
    #   pyproj
    #   requests
    #   sentry-sdk
chardet==4.0.0
    # via requests
click-plugins==1.1.1
    # via fiona
click==7.1.2
    # via fiona
cligj==0.7.1
    # via fiona
django-cors-headers==3.7.0
    # via -r requirements.in
django-environ==0.4.5
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django-parler==2.2
    # via -r requirements.in
django==3.1.7
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-filter
    #   djangorestframework
djangorestframework-gis==0.17
    # via -r requirements.in
djangorestframework==3.12.2
    # via
    #   -r requirements.in
    #   djangorestframework-gis
fiona==1.8.18
    # via geopandas
folium==0.12.1
    # via gtfs-kit
geopandas==0.9.0
    # via gtfs-kit
gtfs-kit==5.0.2
    # via -r requirements.in
idna==2.10
    # via requests
jinja2==2.11.3
    # via
    #   branca
    #   folium
json2html==1.3.0
    # via gtfs-kit
markupsafe==1.1.1
    # via jinja2
munch==2.5.0
    # via fiona
numpy==1.20.1
    # via
    #   folium
    #   pandas
pandas==1.2.3
    # via
    #   geopandas
    #   gtfs-kit
psycopg2==2.8.6
    # via -r requirements.in
pycountry==19.8.18
    # via gtfs-kit
pyproj==3.0.1
    # via geopandas
python-dateutil==2.8.1
    # via pandas
pytz==2021.1
    # via
    #   django
    #   pandas
requests==2.25.1
    # via
    #   folium
    #   gtfs-kit
rtree==0.9.7
    # via gtfs-kit
sentry-sdk==0.20.3
    # via -r requirements.in
shapely==1.7.1
    # via
    #   geopandas
    #   gtfs-kit
six==1.15.0
    # via
    #   fiona
    #   python-dateutil
sqlparse==0.4.1
    # via django
urllib3==1.26.3
    # via
    #   requests
    #   sentry-sdk
utm==0.7.0
    # via gtfs-kit
